Fix `Sf.Graphics.Rect.contains`
This problem was introduced in b4b35b941ede652c088db98e41c844b2ad14e639
when trying to simplify the API. It is possible to remove `access constant`
but if the Convention for the records is set to `C` instead of
`C_Pass_By_Copy`. In that case, the `intersects` functions should also be
changed, so to minimize the API changes, `access constant` for `contains`
has been restored. Moreover, other parts of the API follow this convention.

Added test for these functions.

Added this test and pong to the Makefile used for compiling all the tests.

See issue #11