Improvements for OpenGL example
- Set different colors for each cube side.
- Fix the vertices order so front polygons face the outside and back
  polygons, the inside of the cube.
- Set fill mode for the front and point mode for the back.

My knowledge of OpenGL is not enough to know if the last point is really
necessary or could be implemented in a better way. Ideally, back faces
should not visible at all when another polygon is in front of it. Currently
the vertices are visible as points through the opposite face.